backend/core/communication/serializers.py
from rest_framework import serializers
from core.models import Announcement

class AnnouncementSerializer(serializers.ModelSerializer):
    class Meta:
        model = Announcement
        fields = "__all__"
